data_scrapers/pyproject.toml
[tool.poetry]
name = "data-scrapers"
version = "0.1.0"
description = ""
authors = ["2jun0 <soo28819@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.31.0"
sqlalchemy = "^2.0.23"
mysqlclient = "^2.2.0"
psycopg2 = "^2.9.9"
sqlalchemy-utils = "^0.41.1"
pydantic-settings = "^2.1.0"
apscheduler = "^3.10.4"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
factory-boy = "^3.3.0"
pytest-mock = "^3.12.0"
pytest-factoryboy = "^2.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 119
experimental-string-processing = true
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''